realtime/requirements.txt
# Real-time Trading Infrastructure Requirements

# Core Dependencies
redis>=4.0.0
pyyaml>=6.0
psutil>=5.8.0

# Kafka (optional - will use mock if not available)
kafka-python>=2.0.2

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Dashboard
streamlit>=1.20.0
plotly>=5.0.0

# Database
sqlite3  # Built into Python

# Monitoring and Logging
requests>=2.25.0

# Development and Testing
pytest>=6.0.0
pytest-cov>=2.10.0

# Optional: Advanced features
# numba>=0.53.0  # For GPU acceleration (if available)
# cupy>=9.0.0    # For CUDA support (if available)